Fix branch deletion

+++ bin/gmerged.sh
@@ -63,9 +63,11 @@ if [ -n "$push" ]; then
 fi
 
 if [ "$delete" -eq 1 ]; then
-    echo -e "\n=== DELETEING REMOTE ==="
-    git push --delete origin "$branch"
-fi
+    echo -e "\n=== DELTEING LOCAL ==="
+    git branch --delete "$branch"
 
-echo -e "\n=== DELETING LOCAL ==="
-git branch -D "$branch"
+    if [ -n "$push" ] && [ -n "$(git ls-remote $(git config --get remote.origin.url) history)" ]; then
+        echo -e "\n=== DELETEING REMOTE ==="
+        git push --delete origin "$branch"
+    fi
+fi
